requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
prefect>=2.10.0
locust>=2.15.0
prometheus-fastapi-instrumentator>=6.0.0
slowapi>=0.1.8
python-multipart>=0.0.6
fastf1>=3.0.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
apscheduler>=3.10.0
structlog>=23.1.0
typer>=0.9.0
pyarrow>=12.0.0
dvc>=3.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
xgboost>=2.0.0
optuna>=3.0.0
mlflow>=2.0.0
dagshub>=0.3.0
evidently>=0.4.0
plotly>=5.0.0
scipy>=1.10.0